Remove relative_url_root
Something is not quite right with `relative_url_root` in Rails 6.1. See
https://github.com/rails/rails/issues/42243 and linked issues.

It will be better to mount the application at a relative URL using the
proxy or server we use, instead of within Rails itself. I've added this
to the list of things we should change in the server configuration.